+++ microsoftdefender/utils/constant.py
@@ -2,13 +2,16 @@ actions = [
     "Warn",
     "Block",
     "Audit",
-    "Alert",
-    "AlertAndBlock",
     "BlockAndRemediate",
     "Allowed"
 ]
 
 allow_deletion = {
+    "Yes",
+    "No"
+}
+
+generate_alert = {
     "Yes",
     "No"
 }